+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Add `- unreleased` along the next future version to prevent CI/CD from triggerin
 - Default urls for the `source` field in manifest files can now use `orbit.ip.name` and `orbit.ip.version` variables for variable substitution
 
 ### Bug Fixes
+- Fixes error during install to continue install even though there already exists an ip with same name and version in cache or downloads without `--force` present
 - Squashes bug regarding library for `get` command not defaulting to ip's name when a library is not explicitly present on external ip
 
 ## 0.11.0